educase/requirements.txt
# ── CORE GUI ────────────────────────────────────────────────────────────────
PySide6==6.7.3                  # Qt6 GUI-фреймворк (LGPL)

# ── DATABASE ────────────────────────────────────────────────────────────────
SQLAlchemy==2.0.36              # ORM (async-ready, 2.x API)
alembic==1.14.0                 # Миграции схемы БД

# ── SECURITY ────────────────────────────────────────────────────────────────
bcrypt==4.2.1                   # Хэширование паролей (cost=12)

# ── REPORTS & EXPORT ────────────────────────────────────────────────────────
reportlab==4.2.5                # PDF генерация
openpyxl==3.1.5                 # Excel (.xlsx) экспорт
python-docx==1.1.2              # Word шаблоны (DocumentEditor задания)
Pillow==11.0.0                  # Изображения: resize, convert, validate

# ── ANALYTICS / CHARTS ──────────────────────────────────────────────────────
matplotlib==3.9.3               # Графики (встраивается в PySide6 через FigureCanvas)
numpy==2.1.3                    # Зависимость matplotlib; используется в analytics_service

# ── UTILITIES ───────────────────────────────────────────────────────────────
loguru==0.7.2                   # Структурированное логирование
